appDisplay: Set popup menu side via parameter to AppIcon class
Let's make the "side" argument to the popupMenu() function for app icons
a param to the AppIcon class itself. The side is not that dynamic that it
needs to be defined on calling of the function, instead it can be defined
on creation of the AppIcon as well.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3796>