ci: Simplify run-eslint script
Now that all code conforms with the new style, we can remove all
the tricky bits that compare errors from regular- and legacy
config or limit checks to changed lines from a git diff.

All that is left over the eslint CLI tool is the ability to output
results both as junit for gitlab and plain text for logs without
duplicating the linting, but that's well worth preserving.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/2866>